Extract session codec from params state
